requirements.txt
python-dotenv==0.20.0
RPi.GPIO==0.7.1
wheel==0.37.1
Adafruit_DHT==1.4.0